fix(dapp): decouple dashboard components from mock data module (#1232)
Extract domain types (Transaction, VaultPosition, PortfolioStats, RiskTier,
LoadingState) into lib/types.ts so production components never import from
lib/mock-data. Performance chart now accepts a performanceHistory prop and
renders an empty state when no real data exists. The hardcoded "+12.5%"
mock yield change is removed.
