+++ apps/dapp/frontend/components/dashboard/recent-activity.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { motion } from "framer-motion";
-import { Transaction } from "@/lib/mock-data";
+import type { Transaction } from "@/lib/types";
 import { cn, truncateAddress } from "@/lib/utils";
 import { 
     ArrowUpRight,
