+++ apps/dapp/frontend/components/dashboard/vault-positions-table.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { motion } from "framer-motion";
-import { VaultPosition, RiskTier } from "@/lib/mock-data";
+import type { VaultPosition, RiskTier } from "@/lib/types";
 import { cn } from "@/lib/utils";
 import Link from 'next/link';
 import { ExternalLink, Plus, Minus } from "lucide-react";
